proper enterprise package structure
