--- main/src/org/jetbrains/bsp/bazel/BUILD
@@ -1,10 +1,0 @@
-java_library(
-    name = "bsp",
-    tags = ["maven_coordinates=org.jetbrains.bsp:bazel-bsp:{pom_version}"],
-    visibility = ["//visibility:public"],
-    runtime_deps = [
-        "//main/src/org/jetbrains/bsp/bazel/common",
-        "//main/src/org/jetbrains/bsp/bazel/install",
-        "//main/src/org/jetbrains/bsp/bazel/server",
-    ],
-)
