+++ src/main/java/org/jetbrains/bsp/bazel/BUILD
@@ -1,0 +1,10 @@
+java_library(
+    name = "bsp",
+    tags = ["maven_coordinates=org.jetbrains.bsp:bazel-bsp:{pom_version}"],
+    visibility = ["//visibility:public"],
+    runtime_deps = [
+        "//src/main/java/org/jetbrains/bsp/bazel/common",
+        "//src/main/java/org/jetbrains/bsp/bazel/install",
+        "//src/main/java/org/jetbrains/bsp/bazel/server",
+    ],
+)
