+++ src/main/java/org/jetbrains/bsp/bazel/common/BUILD
