+++ src/main/java/org/jetbrains/bsp/bazel/server/BUILD
@@ -1,0 +1,30 @@
+java_library(
+    name = "server",
+    srcs = glob(["**/*.java"]),
+    visibility = ["//visibility:public"],
+    deps = [
+        "//src/main/java/org/jetbrains/bsp/bazel/common",
+        "@com_google_protobuf//:protobuf_java",
+        "@googleapis//:google_devtools_build_v1_build_events_java_proto",
+        "@googleapis//:google_devtools_build_v1_publish_build_event_java_grpc",
+        "@googleapis//:google_devtools_build_v1_publish_build_event_java_proto",
+        "@googleapis//:google_devtools_build_v1_publish_build_event_proto",
+        "@io_bazel//src/main/java/com/google/devtools/build/lib/buildeventstream/proto:build_event_stream_java_proto",
+        "@io_bazel//src/main/protobuf:analysis_java_proto",
+        "@io_bazel//src/main/protobuf:build_java_proto",
+        "@io_bazel//third_party/grpc:grpc-jar",
+        "@io_bazel_rules_scala//src/protobuf/io/bazel/rules_scala:diagnostics_java_proto",
+        "@maven//:ch_epfl_scala_bsp4j",
+        "@maven//:com_google_code_gson_gson",
+        "@maven//:com_google_guava_guava",
+        "@maven//:org_eclipse_lsp4j_org_eclipse_lsp4j_jsonrpc",
+        "@maven//:org_eclipse_xtext_org_eclipse_xtext_xbase_lib",
+    ],
+)
+
+java_binary(
+    name = "bsp-run",
+    main_class = "org.jetbrains.bsp.bazel.Server",
+    visibility = ["//visibility:public"],
+    runtime_deps = ["//src/main/java/org/jetbrains/bsp/bazel/server"],
+)
